Move MinimumRequiredStorageDeposit from Account to Client
