+++ bindings/nodejs/lib/types/client/bridge/client.ts
@@ -4,7 +4,7 @@
 import type { SecretManagerType } from '../../secret_manager/secret-manager';
 import type { IGenerateAddressesOptions } from '../generate-addresses-options';
 import type { IBuildBlockOptions } from '../build-block-options';
-import type { Block, BlockId, Payload } from '../../block';
+import type { Block, BlockId, Output, Payload } from '../../block';
 import type {
     PreparedTransactionData,
     IBip32Chain,
@@ -423,6 +423,13 @@ export interface __ClearListenersMethod__ {
         topics: string[];
     };
 }
+
+export type __MinimumRequiredStorageDepositMethod__ = {
+    name: 'minimumRequiredStorageDeposit';
+    data: {
+        output: Output;
+    };
+};
 
 export type __RequestFundsFromFaucetMethod__ = {
     name: 'requestFundsFromFaucet';
