+++ bindings/nodejs/lib/types/client/bridge/index.ts
@@ -65,6 +65,7 @@ import type {
     __BuildNftOutputMethod__,
     __ClearListenersMethod__,
     __SignatureUnlockMethod__,
+    __MinimumRequiredStorageDepositMethod__,
     __RequestFundsFromFaucetMethod__,
     __CallPluginRouteMethod__,
 } from './client';
@@ -133,5 +134,6 @@ export type __ClientMethods__ =
     | __BuildFoundryOutputMethod__
     | __BuildNftOutputMethod__
     | __ClearListenersMethod__
+    | __MinimumRequiredStorageDepositMethod__
     | __RequestFundsFromFaucetMethod__
     | __CallPluginRouteMethod__;
