+++ bindings/python/examples/how_tos/nfts/burn_nft.py
@@ -15,10 +15,9 @@ wallet.set_stronghold_password(os.environ["STRONGHOLD_PASSWORD"])
 account = wallet.get_account('Alice')
 
 # Sync account with the node
-response = account.sync()
+balance = account.sync()
 
-# TODO: replace with your own values.
-nftId = "0xf95f4d5344217a2ba19a6c19a47f97d267edf8c4d76a7b8c08072ad35acbebbe"
+nftId = balance.nfts[0]
 
 # Send transaction.
 transaction = account.prepare_burn_nft(nftId).send()
